DEV-974 Replace fixed sleep when updating metadata
Although we get a successful response back when creating a fragment
it doesn't mean it has already been created or is available. As a
result, calling the update metadata request immediately after the
creation might result in a 404. A temporary fix was in place that
would just delay for a couple of seconds. This is replaced by a more
flexible system. It will retry for a couple of times with an
exponential back-off.