Fix logger issues
When initializing a logger based on a ConfigParser, it will load in the
entire config.yml as it uses the logging-level specified in the config.
However it will fail if at that time the env variables are not yet set.
In the context of running the tests, these env variables are patched/set
in conftest.py. Unfortunately, this happens after collecting the tests.

For now, to alleviate this, the logger level in the retry decorator
will be hardcoded to DEBUG so it doesn't need to parse the config.yml.

The logger in the PID service will be removed for now as this is less
crucial.

Also, add missing env variables in conftest.py.