DEV-1031 Retry metadata update when a 403 occurs
In the case of an essence linked, after creating a fragment we
immediately send an update metadata call. This sometimes results in a
"403 - no access", although the same user executes the create fragment
as well as the update metadata calls.

The assumption is that MH first creates the fragment but has not yet
correctly applied the rights. As a fix we will retry sending the
request with an exponential back-off when getting a 403 back.