DEV-1042 check s3_object_key on essence level
It is possible that multiple essenceLinkedEvents come in for an
essence. The fragments share that key, so the query returns multiple
results if events have come in for that essence. Thus, the check will
fail for every event except the first.

The check should be on unique on the essence level.