DEV-2392 Retry when fragment is not found
Although we get a successful response back when creating a fragment
it doesn't mean it has already been created or is available. As a
result, calling the update metadata request immediately after the
creation might result in a 404 or 403. There was a functionality
that would retry in the case this happens. This functionality,
however, was lost with the change to using the MediaHaven library.

Re-add that functionality.